% Auto-generated: do not edit by hand
\name{compare}

\alias{compare}

\title{Compare component}

\description{
This will be used for searching and comparing names
}

\usage{
compare(props=NULL)
}

\arguments{
\item{props}{A list of or a singular dash component, string or number. Props being passed down from parent}
}
