% Auto-generated: do not edit by hand
\name{famename}

\alias{famename}

\title{Famename component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
famename(id=NULL, genderSelect=NULL, nameOutput=NULL, currentPage=NULL, gender=NULL,
ethnicity=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{genderSelect}{Character. Selects gender based on user's selection.}

\item{nameOutput}{Unnamed list. List of names based on gender selected by user.}

\item{currentPage}{Character. Shows current user page.}

\item{gender}{Character. Filter rank datatable by gender.}

\item{ethnicity}{Character. Filter rank datatable by ethnicity.}
}
