% Auto-generated: do not edit by hand
\name{genderSelectContainer}

\alias{genderSelectContainer}

\title{GenderSelectContainer component}

\description{
Allows users to select a gender to return a list of random names.
}

\usage{
genderSelectContainer(props=NULL)
}

\arguments{
\item{props}{A list of or a singular dash component, string or number. Props being passed down from parent}
}
