% Auto-generated: do not edit by hand
\name{style}

\alias{style}

\title{style component}

\description{

}

\usage{
style(selectState=NULL, handleButtonClick=NULL, genderEthnicity=NULL, cssArray=NULL,
last=NULL, ethnicity=NULL)
}

\arguments{
\item{selectState}{Character. }

\item{handleButtonClick}{}

\item{genderEthnicity}{Character. }

\item{cssArray}{Unnamed list. }

\item{last}{Logical. }

\item{ethnicity}{Logical. }
}
