% Auto-generated: do not edit by hand
\name{genderSelectContainer}

\alias{genderSelectContainer}

\title{GenderSelectContainer component}

\description{

}

\usage{
genderSelectContainer(nameOutput=NULL, selectedName=NULL)
}

\arguments{
\item{nameOutput}{Unnamed list. }

\item{selectedName}{Character. }
}
